+++ lib/anthracite.hpp
@@ -1,4 +1,6 @@
 #include "backends/backend.hpp"
+#include "config/config.hpp"
 
 using namespace anthracite;
-int anthracite_main(int argc, char** argv, backends::backend& be);
+
+int anthracite_main(backends::backend& be, config::config& cfg);