+++ lib/socket/socket.cpp
@@ -29,13 +29,22 @@ anthracite_socket::anthracite_socket(int port, int max_queue)
     listen(server_socket, max_queue);
 }
 
-void anthracite_socket::wait_for_conn()
+bool anthracite_socket::wait_for_conn()
 {
     client_ip = "";
-    client_socket = accept(server_socket, reinterpret_cast<struct sockaddr*>(&client_addr), &client_addr_len);
-    std::array<char, INET_ADDRSTRLEN> ip_str { 0 };
-    inet_ntop(AF_INET, &client_addr.sin_addr, ip_str.data(), INET_ADDRSTRLEN);
-    client_ip = std::string(ip_str.data());
+    struct timeval tv = {.tv_sec = 1, .tv_usec = 0};
+    fd_set read_fd;
+    FD_ZERO(&read_fd);
+    FD_SET(server_socket, &read_fd);
+    if (select(server_socket+1, &read_fd, NULL, NULL, &wait_timeout)) {
+        client_socket = accept(server_socket, reinterpret_cast<struct sockaddr*>(&client_addr), &client_addr_len);
+        std::array<char, INET_ADDRSTRLEN> ip_str { 0 };
+        inet_ntop(AF_INET, &client_addr.sin_addr, ip_str.data(), INET_ADDRSTRLEN);
+        client_ip = std::string(ip_str.data());
+        return true;
+    } else {
+        return false;
+    }
 }
 
 const std::string& anthracite_socket::get_client_ip()